+++ src/k8s-configuration/azext_k8s_configuration/_params.py
@@ -137,21 +137,21 @@ def load_arguments(self, _):
             help="File path to known_hosts contents containing public SSH keys required to access private Git instances",
         )
         c.argument(
-            "access_key",
+            "bucket_access_key",
             arg_group="Bucket Auth",
-            help="Access Key Id used to authenticate with the bucket url",
+            help="Access Key ID used to authenticate with the bucket",
         )
         c.argument(
-            "secret_key",
+            "bucket_secret_key",
             arg_group="Bucket Auth",
-            help="Secret Key used to authenticate with the bucket url",
+            help="Secret Key used to authenticate with the bucket",
         )
         c.argument(
             "bucket_name",
-            help="Name of the S3 bucket at the given url to connect to",
+            help="Name of the S3 bucket to sync",
         )
         c.argument(
-            "insecure",
+            "bucket_insecure",
             arg_type=get_three_state_flag(),
             help="Communicate with a bucket without TLS",
         )