+++ src/k8s-configuration/azext_k8s_configuration/providers/FluxConfigurationProvider.py
@@ -144,9 +144,9 @@ def create_config(
     https_ca_cert_file=None,
     known_hosts=None,
     known_hosts_file=None,
-    access_key=None,
-    secret_key=None,
-    insecure=False,
+    bucket_access_key=None,
+    bucket_secret_key=None,
+    bucket_insecure=False,
     suspend=False,
     kustomization=None,
     no_wait=False,
@@ -175,9 +175,9 @@ def create_config(
         https_ca_cert_file=https_ca_cert_file,
         known_hosts=known_hosts,
         known_hosts_file=known_hosts_file,
-        access_key=access_key,
-        secret_key=secret_key,
-        insecure=insecure,
+        bucket_access_key=bucket_access_key,
+        bucket_secret_key=bucket_secret_key,
+        bucket_insecure=bucket_insecure,
     )
 
     # This update func is a generated update function that modifies
@@ -193,7 +193,7 @@ def create_config(
 
     # Get the protected settings and validate the private key value
     protected_settings = get_protected_settings(
-        ssh_private_key, ssh_private_key_file, https_key, secret_key
+        ssh_private_key, ssh_private_key_file, https_key, bucket_secret_key
     )
     if protected_settings and consts.SSH_PRIVATE_KEY_KEY in protected_settings:
         validate_private_key(protected_settings["sshPrivateKey"])
@@ -256,9 +256,9 @@ def update_config(
     https_ca_cert_file=None,
     known_hosts=None,
     known_hosts_file=None,
-    access_key=None,
-    secret_key=None,
-    insecure=None,
+    bucket_access_key=None,
+    bucket_secret_key=None,
+    bucket_insecure=None,
     suspend=None,
     kustomization=None,
     no_wait=False,
@@ -293,9 +293,9 @@ def update_config(
         https_ca_cert_file=https_ca_cert_file,
         known_hosts=known_hosts,
         known_hosts_file=known_hosts_file,
-        access_key=access_key,
-        secret_key=secret_key,
-        insecure=insecure,
+        bucket_access_key=bucket_access_key,
+        bucket_secret_key=bucket_secret_key,
+        bucket_insecure=bucket_insecure,
     )
 
     # This update func is a generated update function that modifies
@@ -324,7 +324,7 @@ def update_config(
 
     # Get the protected settings and validate the private key value
     protected_settings = get_protected_settings(
-        ssh_private_key, ssh_private_key_file, https_key, secret_key
+        ssh_private_key, ssh_private_key_file, https_key, bucket_secret_key
     )
     if protected_settings and consts.SSH_PRIVATE_KEY_KEY in protected_settings:
         validate_private_key(protected_settings["sshPrivateKey"])
@@ -968,15 +968,17 @@ class BucketGenerator(SourceKindGenerator):
         self.bucket_name = kwargs.get("bucket_name")
         self.timeout = kwargs.get("timeout")
         self.sync_interval = kwargs.get("sync_interval")
-        self.access_key = kwargs.get("access_key")
-        self.secret_key = kwargs.get("secret_key")
+        self.bucket_access_key = kwargs.get("bucket_access_key")
+        self.bucket_secret_key = kwargs.get("bucket_secret_key")
         self.local_auth_ref = kwargs.get("local_auth_ref")
-        self.insecure = kwargs.get("insecure")
+        self.bucket_insecure = kwargs.get("bucket_insecure")
 
     def validate(self):
         super().validate_required_params(**self.kwargs)
         validate_bucket_url(self.url)
-        if not ((self.access_key and self.secret_key) or self.local_auth_ref):
+        if not (
+            (self.bucket_access_key and self.bucket_secret_key) or self.local_auth_ref
+        ):
             raise RequiredArgumentMissingError(
                 consts.REQUIRED_BUCKET_VALUES_MISSING_ERROR,
                 consts.REQUIRED_BUCKET_VALUES_MISSING_HELP,
@@ -995,9 +997,9 @@ class BucketGenerator(SourceKindGenerator):
                 bucket_name=self.bucket_name,
                 timeout_in_seconds=parse_duration(self.timeout),
                 sync_interval_in_seconds=parse_duration(self.sync_interval),
-                access_key=self.access_key,
+                access_key=self.bucket_access_key,
                 local_auth_ref=self.local_auth_ref,
-                insecure=self.insecure,
+                insecure=self.bucket_insecure,
             )
             config.source_kind = SourceKindType.BUCKET
             return config
@@ -1018,9 +1020,9 @@ class BucketGenerator(SourceKindGenerator):
                     bucket_name=self.bucket_name,
                     timeout_in_seconds=parse_duration(self.timeout),
                     sync_interval_in_seconds=parse_duration(self.sync_interval),
-                    access_key=self.access_key,
+                    access_key=self.bucket_access_key,
                     local_auth_ref=self.local_auth_ref,
-                    insecure=self.insecure,
+                    insecure=self.bucket_insecure,
                 )
                 if swapped_kind:
                     self.validate()
@@ -1032,7 +1034,7 @@ class BucketGenerator(SourceKindGenerator):
 
 
 def get_protected_settings(
-    ssh_private_key, ssh_private_key_file, https_key, secret_key
+    ssh_private_key, ssh_private_key_file, https_key, bucket_secret_key
 ):
     protected_settings = {}
     ssh_private_key_data = get_data_from_key_or_file(
@@ -1046,8 +1048,8 @@ def get_protected_settings(
     if https_key:
         protected_settings[consts.HTTPS_KEY_KEY] = to_base64(https_key)
 
-    if secret_key:
-        protected_settings[consts.BUCKET_SECRET_KEY_KEY] = to_base64(secret_key)
+    if bucket_secret_key:
+        protected_settings[consts.BUCKET_SECRET_KEY_KEY] = to_base64(bucket_secret_key)
 
     # Return the protected settings dict if there are any values there
     return protected_settings if len(protected_settings) > 0 else None