+++ src/spring-cloud/azext_spring_cloud/_transformers.py
@@ -83,13 +83,13 @@ def transform_spring_cloud_certificate_output(result):
         result = [result]
 
     for item in result:
-        item['Certificate Name'] = item['properties']['keyVaultCertName']
+        item['Certificate Name'] = item['name']
         item['Thumbprint'] = item['properties']['thumbprint']
         item['Activate Date'] = item['properties']['activateDate'].split("T")[0]
-        item['Certificate Version'] = item['properties']['certVersion']
         item['DNS Names'] = item['properties']['dnsNames']
         item['Expiration Date'] = item['properties']['expirationDate'].split("T")[0]
-        item['Key Vault Uri'] = item['properties']['vaultUri']
+        item['Certificate Version'] = item['properties'].get('certVersion', "")
+        item['Key Vault Uri'] = item['properties'].get('vaultUri', "")
 
     return result if is_list else result[0]