+++ src/connectedk8s/HISTORY.rst
@@ -2,10 +2,15 @@
 
 Release History
 ===============
-1.2.3
+1.2.4
 ++++++
 
 * Added private link support
+
+1.2.3
+++++++
+
+* Fetching the tenantID from subscription object instead of graphclient
 
 1.2.2
 ++++++